integrate LocationManagerDF to DependencyFactory, minor change in GpsTracker
